Move Triggers accordion files into a new folder
